requirements.txt
flask
psutil
py-cpuinfo
gputil
